+++ android/src/org/coolreader/crengine/CrdbLibraryMetadataStore.java
@@ -1,0 +1,38 @@
+/*
+ * CoolReader for Android
+ *
+ * This program is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation, either version 2
+ * of the License, or (at your option) any later version.
+ */
+
+package org.coolreader.crengine;
+
+import org.coolreader.db.CRDBService;
+
+import java.util.ArrayList;
+import java.util.Collection;
+
+final class CrdbLibraryMetadataStore implements LibraryMetadataStore {
+	private final CRDBService.LocalBinder database;
+
+	CrdbLibraryMetadataStore(CRDBService.LocalBinder database) {
+		if (database == null)
+			throw new IllegalArgumentException(
+					"database must not be null");
+		this.database = database;
+	}
+
+	@Override
+	public void load(ArrayList<String> pathNames,
+			Scanner.ScanControl control, LoadCallback callback) {
+		database.loadFileInfos(
+				pathNames, control, callback::onLoaded);
+	}
+
+	@Override
+	public void save(Collection<FileInfo> files) {
+		database.saveFileInfos(files);
+	}
+}