+++ android/src/org/coolreader/crengine/LibraryMetadataExtractor.java
@@ -1,0 +1,16 @@
+/*
+ * CoolReader for Android
+ *
+ * This program is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation, either version 2
+ * of the License, or (at your option) any later version.
+ */
+
+package org.coolreader.crengine;
+
+interface LibraryMetadataExtractor {
+	boolean extractProperties(FileInfo file);
+
+	boolean updateFileFingerprint(FileInfo file);
+}